# server/src/generated/routes.ts
/* tslint:disable */
/* eslint-disable */
// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
import { Controller, ValidationService, FieldErrors, ValidateError, TsoaRoute, HttpStatusCodeLiteral, TsoaResponse, fetchMiddlewares } from "@tsoa/runtime"
// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
import { AppointmentsController } from "./../http/controllers/appointments/appointments.controller.js"
import type { RequestHandler, Router } from "express"

// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

const models: TsoaRoute.Models = {
  TCreateContextResponse: {
    dataType: "refAlias",
    type: {
      dataType: "nestedObjectLiteral",
      nestedProperties: {
        form_url: { dataType: "string", required: true },
        cle_ministere_educatif: { dataType: "string", required: true },
        id_rco_formation: { dataType: "string", required: true },
        localite: { dataType: "string", required: true },
        cfd: { dataType: "string", required: true },
        etablissement_formateur_siret: { dataType: "string", required: true },
        code_postal: { dataType: "string", required: true },
        lieu_formation_adresse: { dataType: "string", required: true },
        intitule_long: { dataType: "string", required: true },
        etablissement_formateur_entreprise_raison_sociale: { dataType: "string", required: true },
      },
      validators: {},
    },
  },
  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
  TCreateContextResponseError: {
    dataType: "refAlias",
    type: { dataType: "nestedObjectLiteral", nestedProperties: { error: { dataType: "string", required: true } }, validators: {} },
  },
  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
  TCreateContextBody: {
    dataType: "refAlias",
    type: {
      dataType: "nestedObjectLiteral",
      nestedProperties: {
        referrer: { dataType: "string", required: true },
        trainingHasJob: { dataType: "boolean" },
        idCleMinistereEducatif: { dataType: "string" },
        idActionFormation: { dataType: "string" },
        idRcoFormation: { dataType: "string" },
        idParcoursup: { dataType: "string" },
      },
      validators: {},
    },
  },
  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
}
const validationService = new ValidationService(models)

// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

export function RegisterRoutes(app: Router) {
  // ###########################################################################################################
  //  NOTE: If you do not see routes for all of your controllers in this file, then you might not have informed tsoa of where to look
  //      Please look into the "controllerPathGlobs" config option described in the readme: https://github.com/lukeautry/tsoa
  // ###########################################################################################################
  app.post(
    "/api/appointment-request/context/create",
    ...fetchMiddlewares<RequestHandler>(AppointmentsController),
    ...fetchMiddlewares<RequestHandler>(AppointmentsController.prototype.createContext),

    function AppointmentsController_createContext(request: any, response: any, next: any) {
      const args = {
        body: { in: "body", name: "body", required: true, ref: "TCreateContextBody" },
      }

      // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

      let validatedArgs: any[] = []
      try {
        validatedArgs = getValidatedArgs(args, request, response)

        const controller = new AppointmentsController()

        const promise = controller.createContext.apply(controller, validatedArgs as any)
        promiseHandler(controller, promise, response, undefined, next)
      } catch (err) {
        return next(err)
      }
    }
  )
  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

  function isController(object: any): object is Controller {
    return "getHeaders" in object && "getStatus" in object && "setStatus" in object
  }

  function promiseHandler(controllerObj: any, promise: any, response: any, successStatus: any, next: any) {
    return Promise.resolve(promise)
      .then((data: any) => {
        let statusCode = successStatus
        let headers
        if (isController(controllerObj)) {
          headers = controllerObj.getHeaders()
          statusCode = controllerObj.getStatus() || statusCode
        }

        // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

        returnHandler(response, statusCode, data, headers)
      })
      .catch((error: any) => next(error))
  }

  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

  function returnHandler(response: any, statusCode?: number, data?: any, headers: any = {}) {
    if (response.headersSent) {
      return
    }
    Object.keys(headers).forEach((name: string) => {
      response.set(name, headers[name])
    })
    if (data && typeof data.pipe === "function" && data.readable && typeof data._read === "function") {
      response.status(statusCode || 200)
      data.pipe(response)
    } else if (data !== null && data !== undefined) {
      response.status(statusCode || 200).json(data)
    } else {
      response.status(statusCode || 204).end()
    }
  }

  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

  function responder(response: any): TsoaResponse<HttpStatusCodeLiteral, unknown> {
    return function (status, data, headers) {
      returnHandler(response, status, data, headers)
    }
  }

  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

  function getValidatedArgs(args: any, request: any, response: any): any[] {
    const fieldErrors: FieldErrors = {}
    const values = Object.keys(args).map((key) => {
      const name = args[key].name
      switch (args[key].in) {
        case "request":
          return request
        case "query":
          return validationService.ValidateParam(args[key], request.query[name], name, fieldErrors, undefined, { noImplicitAdditionalProperties: "throw-on-extras" })
        case "queries":
          return validationService.ValidateParam(args[key], request.query, name, fieldErrors, undefined, { noImplicitAdditionalProperties: "throw-on-extras" })
        case "path":
          return validationService.ValidateParam(args[key], request.params[name], name, fieldErrors, undefined, { noImplicitAdditionalProperties: "throw-on-extras" })
        case "header":
          return validationService.ValidateParam(args[key], request.header(name), name, fieldErrors, undefined, { noImplicitAdditionalProperties: "throw-on-extras" })
        case "body":
          return validationService.ValidateParam(args[key], request.body, name, fieldErrors, undefined, { noImplicitAdditionalProperties: "throw-on-extras" })
        case "body-prop":
          return validationService.ValidateParam(args[key], request.body[name], name, fieldErrors, "body.", { noImplicitAdditionalProperties: "throw-on-extras" })
        case "formData":
          if (args[key].dataType === "file") {
            return validationService.ValidateParam(args[key], request.file, name, fieldErrors, undefined, { noImplicitAdditionalProperties: "throw-on-extras" })
          } else if (args[key].dataType === "array" && args[key].array.dataType === "file") {
            return validationService.ValidateParam(args[key], request.files, name, fieldErrors, undefined, { noImplicitAdditionalProperties: "throw-on-extras" })
          } else {
            return validationService.ValidateParam(args[key], request.body[name], name, fieldErrors, undefined, { noImplicitAdditionalProperties: "throw-on-extras" })
          }
        case "res":
          return responder(response)
      }
    })

    if (Object.keys(fieldErrors).length > 0) {
      throw new ValidateError(fieldErrors, "")
    }
    return values
  }

  // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
}

// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa